requirements.txt
Django==3.1.5
django-cors-headers==3.6.0
djangorestframework==3.12.2
-e git+https://github.com/SimpleJWT/django-rest-framework-simplejwt#egg=djangorestframework-simplejwt
gunicorn==20.0.4
PyJWT==2.0.1
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
swagger-ui==0.1.2
whitenoise==5.2.0
